Vue: extract default value function as string
The default value of a property can be defined as function.
If such a function returns for example a random value, it is not helpful
to extract a random value as default value. It would be better to see
the function instead to get an idea of what kind of value will be returned.

Therefore, default value functions are returned as string,
if they do not initialize an empty Array or Object.